% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beckerkristal2010}
\alias{beckerkristal2010}
\title{Becker-Kristal 2010 formant data in long format}
\format{
A tibble.

\describe{
\item{ISO 639-3}{Language ISO 639-3 code.}
\item{language}{Language name.}
\item{variety}{Language variety.}
\item{genetics}{Philogenetic relation.}
\item{source}{Source of formant values.}
\item{n_speakers}{Number of speakers from which values were taken.}
\item{method}{Method of speech elicitation.}
\item{quantity}{Vowel phonological quantity.}
\item{harmony}{Does the language have vowel harmony?}
\item{n_vowels}{Total number of vowels in the language (excluding length/nasalisation contrasts).}
\item{structure}{Vowel system structure as proposed in Becker-Kristal 2010.}
\item{n_peripheral}{Number of peripheral vowels.}
\item{symmetry}{Symmetry of the vowel system.}
\item{n_nonperipheral}{Number of non-peripheral vowels.}
\item{OS}{Original symbol.}
\item{PS}{Alternative symbol.}
\item{F1, F4}{Values of first, second, third and fourth formants.}
\item{vowel_num}{Numeric ID of the vowel within each language.}
}
}
\source{
\url{http://phonetics.linguistics.ucla.edu/facilities/databases/databases.html}
}
\usage{
beckerkristal2010
}
\description{
The Becker-Kristal 2010 formant data in a long format.
}
\keyword{datasets}
