% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beckerkristal2010_wide}
\alias{beckerkristal2010_wide}
\title{Becker-Kristal 2010 formant data in wide format}
\format{
A tibble.

\describe{
\item{ISO 639-3}{Language ISO 639-3 code.}
\item{Language}{Language name.}
}
}
\source{
\url{http://phonetics.linguistics.ucla.edu/facilities/databases/databases.html}
}
\usage{
beckerkristal2010_wide
}
\description{
The Becker-Kristal 2010 formant data in a wide format.
}
\keyword{datasets}
